# blade-service/blade-bg-admin/src/main/java/org/springblade/bgadmin/modules/sys/controller/SysRoleController.java
/**
 * Copyright 2018 首页 http://www.finepetro.com
 * <p>
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 * <p>
 * http://www.apache.org/licenses/LICENSE-2.0
 * <p>
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 */

package org.springblade.bgadmin.modules.sys.controller;

import com.baomidou.mybatisplus.core.metadata.IPage;
import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
import org.springblade.bgadmin.modules.sys.entity.SysRoleEntity;
import org.springblade.bgadmin.modules.sys.form.BaseForm;
import org.springblade.bgadmin.modules.sys.service.SysRoleDeptService;
import org.springblade.bgadmin.modules.sys.service.SysRoleMenuService;
import org.springblade.bgadmin.modules.sys.service.SysRoleService;
import org.springblade.common.utils.R;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

/**
 * 角色管理
 * 
 * @author chenshun
 * @email sunlightcs@gmail.com
 * @date 2016年11月8日 下午2:18:33
 */
@RestController
@RequestMapping("/sys/role")
public class SysRoleController extends AbstractController {
	@Autowired
	private SysRoleService sysRoleService;
	@Autowired
	private SysRoleMenuService sysRoleMenuService;
	@Autowired
	private SysRoleDeptService sysRoleDeptService;
	
	/**
	 * 角色列表
	 */
	@RequestMapping("/list")
	//@RequiresPermissions("sys:role:list")
	public R list(@RequestBody BaseForm baseForm){
	//	PageUtils page = sysRoleService.queryPage(params);

		IPage page = new Page(baseForm.getPage(),baseForm.getSize());
		return R.ok().put("page", sysRoleService.page(page));
	}
	
	/**
	 * 角色列表
	 */
	@RequestMapping("/select")
	//@RequiresPermissions("sys:role:select")
	public R select(){
		List<SysRoleEntity> list = sysRoleService.list(null);
		
		return R.ok().put("list", list);
	}
	
	/**
	 * 角色信息
	 */
	@RequestMapping("/info/{roleId}")
	//@RequiresPermissions("sys:role:info")
	public R info(@PathVariable("roleId") Long roleId){
		SysRoleEntity role = sysRoleService.getById(roleId);
		
		//查询角色对应的菜单
		List<Long> menuIdList = sysRoleMenuService.queryMenuIdList(roleId);
		role.setMenuIdList(menuIdList);

		//查询角色对应的部门
		List<Long> deptIdList = sysRoleDeptService.queryDeptIdList(new Long[]{roleId});
		role.setDeptIdList(deptIdList);
		
		return R.ok().put("role", role);
	}
	
	/**
	 * 保存角色
	 */
//	@SysLog("保存角色")
	@RequestMapping("/save")
	//@RequiresPermissions("sys:role:save")
	public R save(@RequestBody SysRoleEntity role){
		//ValidatorUtils.validateEntity(role);
		
		sysRoleService.save(role);
		
		return R.ok();
	}
	
	/**
	 * 修改角色
	 */
//	@SysLog("修改角色")
	@RequestMapping("/update")
	//@RequiresPermissions("sys:role:update")
	public R update(@RequestBody SysRoleEntity role){
		//ValidatorUtils.validateEntity(role);
		
	//	sysRoleService.updateById(role);
		sysRoleService.update(role);
		
		return R.ok();
	}
	
	/**
	 * 删除角色
	 */
//	@SysLog("删除角色")
	@RequestMapping("/delete")
	//@RequiresPermissions("sys:role:delete")
	public R delete(@RequestBody Long[] roleIds){
		sysRoleService.deleteBatch(roleIds);
		
		return R.ok();
	}
}